feat(cli): achieve CLI ↔ Framework API parity (ADR-0066)
- Implemented 11 missing subcommands: delete, get, update, disassociate, associations, traverse, path, probe-filtered, stats, metrics, watch.
- Exposed missing framework APIs: `incoming_associations`, `reset_metrics`.
- Added `Serialize` to `FrameworkMetricsSnapshot` and `FrameworkStats` for JSON output.
- Extracted `FrameworkMetrics` to `src/framework_metrics.rs` to keep `src/framework.rs` under the 500 LOC gate.
- Moved non-core `Singularity` methods to `src/singularity_ext.rs` to keep `src/singularity.rs` under the 500 LOC gate.
- Added comprehensive `CLI Reference` page to the book.
- Resolved CLI flag conflict for 'traverse' command.
- Added and verified integration tests in `tests/cli_parity.rs`.
- Fixed clippy warnings (needless borrows, items after test module) and type annotation issues.
- Updated `llms.txt` and `llms-full.txt` via `gen-llms-txt.sh`.
