+++ scripts/split_singularity.py
@@ -1,0 +1,30 @@
+import os
+
+with open('src/singularity.rs', 'r') as f:
+    lines = f.readlines()
+
+header = lines[:7]
+# Keep imports and core struct in singularity.rs
+# Find end of struct Singularity
+struct_end = 0
+for i, line in enumerate(lines):
+    if line.startswith('pub struct Singularity {'):
+        for j in range(i, len(lines)):
+            if lines[j].strip() == '}':
+                struct_end = j + 1
+                break
+        break
+
+struct_code = lines[:struct_end]
+
+# Extract implementation methods to separate modules
+impl_start = 0
+for i in range(struct_end, len(lines)):
+    if lines[i].startswith('impl Singularity {'):
+        impl_start = i
+        break
+
+# We will move all associations-related methods and some others to singularity_ext.rs
+# or a new file.
+# For now let's just move some large chunks to singularity_ext.rs if it's already there
+# or create singularity_ops.rs
