+++ src/singularity.rs
@@ -326,40 +326,6 @@ impl Singularity {
         HVec10240::bundle(&vectors)
     }
 
-    pub fn concept_ids(&self) -> Vec<String> {
-        self.concepts.keys().cloned().collect()
-    }
-
-    pub fn all_concepts(&self) -> Vec<Concept> {
-        self.concepts.values().cloned().collect()
-    }
-
-    pub fn all_associations(&self) -> Vec<(String, String, f32)> {
-        let mut output = Vec::new();
-        for (from, links) in &self.associations {
-            for (to, strength) in links {
-                output.push((from.clone(), to.clone(), *strength));
-            }
-        }
-        output
-    }
-
-    pub fn len(&self) -> usize {
-        self.concepts.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.concepts.is_empty()
-    }
-
-    pub fn cache_metrics_snapshot(&self) -> CacheMetricsSnapshot {
-        self.cache_metrics.snapshot()
-    }
-
-    pub fn reset_cache_metrics(&self) {
-        self.cache_metrics.reset();
-    }
-
     fn evict_oldest_if_needed(&mut self) {
         let Some(limit) = self.config.max_concepts else {
             return;
